Add files via upload
Comprehension - Decision Tree Classification in Python

Let’s consider the heart disease data set that we discussed in the earlier segment. The data lists various tests that were conducted on patients along with some other details of the patients. Now, given the test results and other attributes, suppose you want to predict whether a person has a heart disease or not.

Please note that:

Heart disease = 0 means that the person does not have heart disease.
Heart disease = 1 means that the person has heart disease.

 

sex = 0 means that the person is female.

sex = 1 means that the person is male.